+++ lib/openapi_parameters/converters.rb
@@ -1,0 +1,42 @@
+# frozen_string_literal: true
+
+module OpenapiParameters
+  module Converters
+    @converters = {}
+
+    class << self
+      attr_reader :converters
+
+      def register(type, converter)
+        converters[type] = converter
+      end
+
+      def [](schema)
+        type = schema && schema['type']
+        converters.fetch(type) do
+          ->(value) { Converter.convert(value, schema) }
+        end
+      end
+    end
+
+    register('integer', lambda do |value|
+      Integer(value, 10)
+    rescue StandardError
+      value
+    end)
+
+    register('number', lambda do |value|
+      Float(value)
+    rescue StandardError
+      value
+    end)
+
+    register('boolean', lambda do |value|
+      if value == 'true'
+        true
+      else
+        value == 'false' ? false : value
+      end
+    end)
+  end
+end