+++ lib/openapi_parameters/header.rb
@@ -17,7 +17,7 @@ module OpenapiParameters
 
         result[parameter.name] = catch :skip do
           value = Unpacker.unpack_value(parameter, headers[parameter.name])
-          @convert ? Converter.convert(value, parameter.schema) : value
+          @convert ? parameter.convert(value) : value
         end
       end
     end