+++ lib/openapi_parameters/path.rb
@@ -21,7 +21,7 @@ module OpenapiParameters
 
         result[parameter.name] = catch :skip do
           value = Unpacker.unpack_value(parameter, path_params[parameter.name])
-          @convert ? Converter.convert(value, parameter.schema) : value
+          @convert ? parameter.convert(value) : value
         end
       end
     end